Separate UnitUtil (new classes commit)
